Split StandardStreamLogger into platform TUs
Match AppRuntime: shared API + one JSRUNTIMEHOST_PLATFORM implementation.
Windows (Win32/UWP), Android, Apple, and Unix each get their own source file;
Android/Apple share the POSIX tee body via an .inl. Also fix IsStarted() doc
comment indentation.

Copilot-Session: 88569c10-a7ff-4373-9a58-afa9c68b8c09
