+++ Core/Foundation/CMakeLists.txt
@@ -5,12 +5,31 @@ set(SOURCES
     "Include/Babylon/StandardStreamLogger.h"
     "Source/DebugTrace.cpp"
     "Source/PerfTrace.cpp"
-    "Source/StandardStreamLogger.cpp")
+    "Source/StandardStreamLogger.cpp"
+    "Source/StandardStreamLoggerPlatform.h")
+
+# Match AppRuntime: shared API TU + one platform TU selected by JSRUNTIMEHOST_PLATFORM.
+# Windows (Win32/UWP) and POSIX (Android/Apple) keep their own redirect implementations;
+# plain Unix is a successful no-op because stdout/stderr already reach the environment.
+if(JSRUNTIMEHOST_PLATFORM STREQUAL "Win32" OR JSRUNTIMEHOST_PLATFORM STREQUAL "UWP")
+    list(APPEND SOURCES "Source/StandardStreamLogger_Windows.cpp")
+elseif(JSRUNTIMEHOST_PLATFORM STREQUAL "Android")
+    list(APPEND SOURCES
+        "Source/StandardStreamLogger_Android.cpp"
+        "Source/StandardStreamLogger_Posix.inl")
+elseif(JSRUNTIMEHOST_PLATFORM STREQUAL "iOS" OR JSRUNTIMEHOST_PLATFORM STREQUAL "macOS")
+    list(APPEND SOURCES
+        "Source/StandardStreamLogger_Apple.cpp"
+        "Source/StandardStreamLogger_Posix.inl")
+else()
+    list(APPEND SOURCES "Source/StandardStreamLogger_Unix.cpp")
+endif()
 
 add_library(Foundation ${SOURCES})
 
 target_include_directories(Foundation
     PRIVATE "Include/Babylon"
+    PRIVATE "Source"
     INTERFACE "Include")
 
 target_link_libraries(Foundation
